requirements.txt
pytorch-lightning
hydra-core
weaver-pytorch-rnx0dvmdxk
